integrate all sensors
